+++ digital-booking-backend/src/main/java/com/grupo9/digitalBooking/music/model/service/CategoryService.java
@@ -1,6 +1,7 @@
 package com.grupo9.digitalBooking.music.model.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.grupo9.digitalBooking.music.model.DTO.RolDTO;
 import com.grupo9.digitalBooking.music.model.entities.Category;
 import com.grupo9.digitalBooking.music.model.DTO.CategoryDTO;
 import com.grupo9.digitalBooking.music.model.repository.ICategory;
@@ -12,6 +13,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import java.util.logging.Logger;
 
 @Service
 public class CategoryService implements ICategoryService {
@@ -19,12 +21,30 @@ public class CategoryService implements ICategoryService {
     @Autowired
     private ICategory categoryRepository;
 
+    private static final Logger LOGGER = Logger.getLogger(String.valueOf(RolService.class));
+
     @Autowired
     ObjectMapper mapper;
 
+    private CategoryDTO saveCategory(CategoryDTO categoryDTO){
+        Category category =mapper.convertValue(categoryDTO, Category.class);
+        CategoryDTO result = mapper.convertValue(categoryRepository.save(category), CategoryDTO.class);
+        return result;
+    }
+
+    private Boolean existById(Long id) {
+        return categoryRepository.findById(id).isPresent();
+    }
+
     @Override
-    public void createCategory(CategoryDTO categoryDTO) {
-        saveCategory(categoryDTO);
+    public CategoryDTO createCategory(CategoryDTO categoryDTO) {
+        CategoryDTO response = null;
+        Boolean existCategory = categoryRepository.findByName(categoryDTO.getName()).isPresent();
+        if(!existCategory) {
+            response = saveCategory(categoryDTO);
+        }
+        LOGGER.info("respuesta: " + response);
+        return response;
     }
 
     @Override
@@ -37,18 +57,30 @@ public class CategoryService implements ICategoryService {
         return categoryDTO;
     }
 
-    private void saveCategory(CategoryDTO categoryDTO){
-        Category category =mapper.convertValue(categoryDTO, Category.class);
-        categoryRepository.save(category);
-    }
+
     @Override
-    public void modifyCategory(CategoryDTO categoryDTO) {
-        saveCategory(categoryDTO);
+    public CategoryDTO modifyCategory(CategoryDTO categoryDTO) {
+
+        CategoryDTO response = null;
+        Boolean validateCategory = existById(categoryDTO.getId());
+
+        if(validateCategory) {
+            response = saveCategory(categoryDTO);
+        }
+        LOGGER.info("response: " + response);
+        return response;
     }
 
     @Override
-    public void removeCategory(Long id) {
-        categoryRepository.deleteById(id);
+    public Boolean removeCategory(Long id) {
+        Boolean response = false;
+        Boolean exist = existById(id);
+        if(exist) {
+            categoryRepository.deleteById(id);
+            response = true;
+        }
+        LOGGER.info("response: " + response);
+        return response;
     }
 
     @Override
